3 issues fixed

+++ cmd/topic/topic_describe.go
@@ -30,7 +30,7 @@ const (
 
 type DescribeArgs struct {
 	Name      string
-	Offset    int32
+	Offset    int64
 	Limit     int32
 	Partition int32
 	File      string
@@ -48,7 +48,7 @@ func DescribeCommand(deps *di.Deps) *cobra.Command {
 	cmd.Flags().StringVar(&args.Name, Name, "", NameUsage)
 	cmd.MarkFlagRequired(Name) //nolint:errcheck
 
-	cmd.Flags().Int32Var(&args.Offset, Offset, -1, OffsetUsage)
+	cmd.Flags().Int64Var(&args.Offset, Offset, -1, OffsetUsage)
 	cmd.Flags().Int32Var(&args.Limit, Limit, DefaultLimit, LimitUsage)
 	cmd.Flags().Int32Var(&args.Partition, Partition, -1, PartitionUsage)
 	common.AddFileFlagVar(cmd, &args.File, FileUsage)
@@ -56,7 +56,7 @@ func DescribeCommand(deps *di.Deps) *cobra.Command {
 	return cmd
 }
 
-func argValueError(name string, value int32) common.CLIError {
+func argValueError(name string, value int64) common.CLIError {
 	return common.NewCLIArgParseError(fmt.Errorf("invalid value for argument '%s' specified: %d", name, value))
 }
 
@@ -85,7 +85,7 @@ func RunDescribeCommand(args *DescribeArgs) di.CmdWithApiFn {
 			return argValueError(Offset, args.Offset)
 		}
 		if args.Limit < 1 {
-			return argValueError(Limit, args.Limit)
+			return argValueError(Limit, int64(args.Limit))
 		}
 
 		request := api.TopicApi.Describe(cli.Context, args.Name)
